backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
httpx==0.28.1
psycopg[binary]==3.2.13
websockets==15.0.1
